Add cloudinary integration
WARNING: this commit of the app doesn't load image thumbnails, but the
full-sized images. When I tested it out, it used up all my memory and I had to
manually reboot my machine.

About the script:

It turns out that cloudinary automatically appends 6 random letters to the
end of the filename when you upload a file (so e.g. "test.jpg" would be
uploaded as "test_gqktwm.jpg"). This meant that the filenames in the database
didn't match the filenames on cloudinary, which broke all my images. So I
opted to fix it on the cloudinary side by writing this script, which uses the
cloudinary API to bulk rename the image files used by the app, removing the 6
letters appended by cloudinary.