Fix gallery page re-rendering whenever the route changes
I.e. if you go from the gallery view to the detail view and then back to the
gallery view, the component appears to be destroyed and then re-created, meaning
all the data has to be re-fetched and the info popup reappears. I'm not sure why
destroying the component completely is the default behavior, but I've found that
Vue provides a way to preserve the components via the <keep-alive> tag, so
that's what I'm using.